add masonry
the next step will be to make the widths of the masonry tiles uniform.
it makes for a better look. also, i need to remove the container
surrounding the boxes. but i have to make sure that there is only one
parent element in the component. and i need to make the masonry box the
whole width of the window... more or less.